상대방을 공격 할 수 있게 구현

+++ Assets/02.Scripts/Character/CharacterCanvasAbility.cs
@@ -22,7 +22,7 @@ public class CharacterCanvasAbility : CharacterAbility
             transform.forward = Camera.main.transform.forward;
         }
 
-        HpSliderUI.value = _owner.Stat.Health / _owner.Stat.MaxHealth;
+        HpSliderUI.value = (float)_owner.Stat.Health / _owner.Stat.MaxHealth;
         StaminaSliderUI.value = _owner.Stat.Stamina / _owner.Stat.MaxStamina;
     }
 }
